requirements.txt
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-mongoengine==1.0.0
flask-talisman==1.0.0
Flask-WTF==0.14.3
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
filemagic==1.6
mongoengine==0.20.0
openai==0.26.5
packaging
pluggy
py==1.9.0
pycparser==2.20
pymongo==3.11.0
pyparsing==2.4.7
pytest==7.2.1
python-dotenv==0.15.0
requests
six==1.15.0
toml
urllib3==1.25.11
Werkzeug
WTForms==2.3.3
